Resolved merge conflicts after refactoring location
